move registry to app router
