+++ apps/web/app/api/registry/entries/route.ts
@@ -1,12 +1,16 @@
-import { NextApiRequest, NextApiResponse } from 'next';
+import { NextRequest, NextResponse } from 'next/server';
 import { vercelDb as db } from 'apps/web/src/utils/datastores/rds';
 import { kv } from 'apps/web/src/utils/datastores/kv';
 import { logger } from 'apps/web/src/utils/logger';
-import { withTimeout } from 'apps/web/pages/api/decorators';
+import { withTimeout } from 'apps/web/app/api/decorators';
 
-const pageKey = 'api.ocs_registry.entries';
-async function handler(req: NextApiRequest, res: NextApiResponse) {
-  const { page = '1', limit = '10', category, curation } = req.query;
+const PAGE_KEY = 'api.ocs_registry.entries';
+
+async function handler(req: NextRequest) {
+  const category = req.nextUrl.searchParams.get('category');
+  const curation = req.nextUrl.searchParams.get('curation');
+  const page = req.nextUrl.searchParams.get('page') ?? '1';
+  const limit = req.nextUrl.searchParams.get('limit') ?? '10';
 
   const pageNum = parseInt(page as string, 10);
   const limitNum = parseInt(limit as string, 10);
@@ -49,13 +53,16 @@ async function handler(req: NextApiRequest, res: NextApiResponse) {
   };
 
   try {
-    await kv.incr(`stat:requests.${pageKey}`);
+    await kv.incr(`stat:requests.${PAGE_KEY}`);
   } catch (error) {
     logger.error('error getting registry entries', error);
   }
   // Set caching headers
-  res.setHeader('Cache-Control', 's-maxage=300, stale-while-revalidate');
-  res.status(200).json(response);
+  return NextResponse.json(response, {
+    headers: {
+      'Cache-Control': 's-maxage=300, stale-while-revalidate',
+    },
+  });
 }
 
-export default withTimeout(handler);
+export const GET = withTimeout(handler);
